+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTest.php
@@ -36,6 +36,16 @@ use Symfony\Component\DependencyInjection\ParameterBag\ParameterBag;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpKernel\DependencyInjection\LoggerPass;
+use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
+use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\HttpKernel\EventListener\KernelControllerListenerInterface;
+use Symfony\Component\HttpKernel\EventListener\KernelExceptionListenerInterface;
+use Symfony\Component\HttpKernel\EventListener\KernelFinishRequestListenerInterface;
+use Symfony\Component\HttpKernel\EventListener\KernelRequestListenerInterface;
+use Symfony\Component\HttpKernel\EventListener\KernelResponseListenerInterface;
+use Symfony\Component\HttpKernel\EventListener\KernelTerminateListenerInterface;
+use Symfony\Component\HttpKernel\EventListener\KernelViewListenerInterface;
+use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Messenger\Tests\Fixtures\SecondMessage;
 use Symfony\Component\Messenger\Transport\TransportFactory;
@@ -1239,6 +1249,58 @@ abstract class FrameworkExtensionTest extends TestCase
         $this->assertEquals($expected, array_keys($container->getDefinition('session_listener')->getArgument(0)->getValues()));
     }
 
+    public function testKernelListenersRegistration()
+    {
+        $mapping = array(
+            KernelControllerListenerInterface::class => KernelEvents::CONTROLLER,
+            KernelExceptionListenerInterface::class => KernelEvents::EXCEPTION,
+            KernelFinishRequestListenerInterface::class => KernelEvents::FINISH_REQUEST,
+            KernelRequestListenerInterface::class => KernelEvents::REQUEST,
+            KernelResponseListenerInterface::class => KernelEvents::RESPONSE,
+            KernelTerminateListenerInterface::class => KernelEvents::TERMINATE,
+            KernelViewListenerInterface::class => KernelEvents::VIEW,
+        );
+
+        $container = $this->createContainerFromFile('full', array(), true, false);
+        $container->addCompilerPass(new ResolveInstanceofConditionalsPass());
+
+        $container->register('multiple_events_listener', MyMultipleEventListener::class)
+            ->setAutoconfigured(true);
+
+        foreach ($mapping as $class => $event) {
+            $container->register("listener_of.$event", $class)
+                ->setAutoconfigured(true);
+        }
+        $container->compile();
+
+        foreach ($mapping as $class => $event) {
+            $definition = $container->findDefinition("listener_of.$event");
+
+            $this->assertTrue(
+                $definition->hasTag('kernel.event_listener'),
+                'Definition should have tag "kernel.event_listener".'
+            );
+
+            $this->assertSame(
+                array(array('event' => $event)),
+                $definition->getTag('kernel.event_listener')
+            );
+        }
+
+        $definition = $container->findDefinition('multiple_events_listener');
+
+        $this->assertTrue(
+            $definition->hasTag('kernel.event_listener'),
+            'Definition should have tag "kernel.event_listener".'
+        );
+
+        $this->assertSame(
+            array(array('event' => 'kernel.response'), array('event' => 'kernel.request')),
+            $definition->getTag('kernel.event_listener'),
+            'Should contain 2 events in the tag definition'
+        );
+    }
+
     protected function createContainer(array $data = array())
     {
         return new ContainerBuilder(new ParameterBag(array_merge(array(
@@ -1368,6 +1430,28 @@ abstract class FrameworkExtensionTest extends TestCase
             default:
                 $this->fail('Unresolved adapter: '.$adapter);
         }
+    }
+}
+
+/**
+ * @internal
+ */
+class MyMultipleEventListener implements KernelRequestListenerInterface, KernelResponseListenerInterface
+{
+    /**
+     * @param GetResponseEvent $event
+     */
+    public function onKernelRequest(GetResponseEvent $event): void
+    {
+        // Do something
+    }
+
+    /**
+     * @param FilterResponseEvent $event
+     */
+    public function onKernelResponse(FilterResponseEvent $event): void
+    {
+        // Do something else
     }
 }